Fixed odd hang when parameter was changed.
We were seeing the following bug:
In client-server mode, when one created a sphere and showed it and then changed
a parameter and hit apply, it would hang. This was because when both processes
were doing vtkPVRenderView::Render(), when a representation was updated, the
client side had code to fetch data-information for some arbitrary panel. This
was happening while the server was still in it's Render() resulting in
mismatched tag errors. Fixed by separating the Update of all representations and
then Render() into two separate calls.