Mark inputs modified when view time changes.
In previous implementation, when a view's time changed, the representations
marked their chain on inputs dirty. This was essential so that the
representation would re-execute. This is not necessary anymore, the
vtkPVDataRepresentation can update itself automatically when time changes.
However, there no indicate to the upstream source proxies that their data
information is invalid. And for that  we still need this hack to mark all inputs
modified. We need to address data-information invalidate a bit more cleanly and
we can git rid of this hack.