Fixed bounds computations.
vtkWidgetRepresentation was reporting bogus bounds on processes where the render
window wasn't initialized. That was resulting in ParaView gathering incorrect
bounds when remote-rendering was disabled in client-server mode. Fixed that by
avoiding gathering of  bounds on  processes where the window is not valid.

Also marked vtkPVCenterAxesActor so that it's bounds are not used in bounds
computation using the vtkActor API.