Refactoring proxy creation/initialization logic.
In order to increase code reuse between Qt and Python based ParaView
applications, this commit starts the effort to add  a controller logic
at the ServerManager layer. New classes, vtkSMParaViewPipelineController
and vtkSMTransferFunctionManager, are part of this new infrastructure
that handles creation and initialization of various types of ParaView
proxies. The code is proxy type depedent, thus using different
initialization code based on whether the proxy is a representation,
filter, transfer function etc. as needed.

As a consequence all the initialization logic from pqProxy subclasses
becomes obsolete and no longer necessary, hence we remove that (and some
clases that are no longer useful).

This commit also moves transfer function management to ServerManager (as
vtkSMTransferFunctionManager). pqPQLookupTableManager and
pqLookupTableManager classes are removed since they are no longer
relevant. vtkSMTransferFunctionManager no longer depends on number of
components in the array. That caused an unnecessary need to update data to
determine how many components are going to be there. It's also easier for users
to realize that all arrays with same name get the same lut.

Change-Id: I5de5c4b0107a4a417d13fa4f470c51cd394b2cee