Fix ColorArrayName domain issue with initial repr is Outline.
When creating a Wavelet source, for example, the domain would not update
to include cellNormals array even after we switch representation type to
"Surface" from "Outline". This was because the SurfaceRepresentation
proxy's data information was never invalidated when representation type
was changed. It makes sense to invalidate subproxies when properties on
the parent-proxy are modified. This fixes the cellNormals issue too.

Change-Id: I04e6a824d61ae6bacb98017546ddf653537c25d9