Update state version control to use pugixml and xpath.
Refactoring XML state version control to use pugixml with xpaths to make
it easier to process the XML state. The vtkSMStateVersionController code
is a little twisted. It converts vtkPVXMLElement to string for
processing using pugixml and then back to vtkPVXMLElement. This is only
a temporary fix. We need to cleanup the API to avoid this multiple
parsing and conversions.

Removing compatibility with states older than ParaView 4.0.1.

Adding support for handling ColorArrayName/ColorAttributeType changes.

Change-Id: I0a4ecb12d85a1eb47cf6159b8bd56587bf31d4fd