Clicking on documentation labels toggles checkbox
When documentation is used for labels, it was displayed with a
QLabel. For int vector properties with boolean domain, the checkbox
should be toggled with the documentation text is clicked. However,
QLabel's do not emit a signal when clicked. The pqLabel subclass of
QLabel introduced in this commit does emit a clicked() signal when the
text it displays is clicked. Linking this to the click() slot in the
QCheckBox for the boolean domain gets the standard QCheckBox behavior.

Change-Id: I7a257689f167bdadf25083d1ecc0567770f37110