Fixed PropertyConversion tests.
Fixed and documented what FieldAssociation and DomainAssociation stand
for on vtkSMArrayListDomain and fixed their usage which was causing test
issues.

Also fixed vtkSMArrayListDomain to pick the first array by default which
is needed for most array list domain properties except when used for
scalar coloring. So added mechanism in vtkSMRepresentedArrayListDomain
to override that behavior explicitly.

Change-Id: I2af6f1bcf22cd034c644d7e90088fb382edb024b