Fixed issues with pvcs.UndoRedo3 test.
The problem was that the "TimeSource" property's value was not recoded
in the undo-stack. As a result, if one undid the creation of a source,
the TimeSources property would still refer to the proxy that no longer
exists. This was causing it to leak. As a result, there was deadlock
in the test when running in parallel since the reader won't call
RequestInformation on all processes.

Fixed by ensuring that the time keeper it tracked by the undo-redo stack
builder. Since we don't want time related properties to be tracked,
explicitly skipping those from the undo-redo tracking.

Had to fix reference counting in vtkSMTimeKeeper to avoid segfaults when
the proxy gets released prematurely when being removed from the
"TimeSources" property.

Change-Id: I851074eb99b08a6a5918f27abd875114cf02b8c0